Time: 4 ms (6.03%), Space: 19.6 MB (74.62%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -18,7 +18,7 @@ public:
         queue<pair<TreeNode*, long long>> q;
         q.push({root,0});
         while(!q.empty()) {
-            long long size = q.size();
+            int size = q.size();
             long long min = q.front().second;
             long long first,last;
             for(int i=0; i<size; i++) {
